requirements.txt
spacy
python-dotenv